refactor(header, dropdownmenu, headerrightblock): made separate components for headerRightBlock
